requirements.txt
PyAutoGUI==0.9.52
Pillow==8.1.0
opencv-python==4.5.1.48
